pyproject.toml
[project]
name = "zenith"
version = "0.1.0"
description = "Zenith Is A CLI-Based AI Coding Agent That Transforms Natural Language Into Efficient, Production-Ready Code!"
readme = "readme.md"
requires-python = ">=3.13"
authors = [
    {name = "Rohit Vilas Ingole", email = "rohit.vilas.ingole@gmail.com"}
]
license = {file = "license"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
urls = {Homepage = "https://github.com/datarohit/zenith"}
dependencies = [
    "rich-cli>=1.8.1",
    "typer>=0.17.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
]

[build-system]
requires = ["setuptools>=80.9.0", "wheel>=0.45.1"]

[project.scripts]
zenith = "zenith.main:run"

[tool.setuptools]
packages = ["zenith"]

[tool.pytest.ini_options]
minversion = "8.4.2"
python_files = [
    "tests.py",
    "test_*.py",
]

[tool.coverage.run]
include = ["zenith/**"]
omit = ["*/tests/*"]

[tool.ruff]
target-version = "py313"
line-length = 120
extend-exclude = [
    "*/tests/*",
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    'ISC',
    "ICN",
    "G",
    'INP',
    'PIE',
    "T20",
    'PYI',
    'PT',
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TC",
    "INT",
    "PTH",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "PERF",
    "RUF",
]
ignore = []
extend-unsafe-fixes = []

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.per-file-ignores]
"*/tests/*" = []
